Replace "no max" with null (#1359)
Changes the resource reward max system to use `null` to indicate no max.

We already were handling "no max provided = no limit" in the cpp code.
This pushes up the option to provide no default -- either implicitly by
not providing it, or explicitly by providing null (useful to overwrite a
different value, but also just to be explicit).

Also changes to cpp code to use uint8_t for maxes instead of floats.
ints would be fine too -- but floats is wrong, given that the limits are
on amount of resource vs amount of reward.

Added a unit test to test.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210735491520542)

---------

+++ mettagrid/src/metta/mettagrid/objects/agent.hpp
@@ -21,7 +21,7 @@ struct AgentConfig : public GridObjectConfig {
               float action_failure_penalty,
               const std::map<InventoryItem, uint8_t>& resource_limits,
               const std::map<InventoryItem, float>& resource_rewards,
-              const std::map<InventoryItem, float>& resource_reward_max,
+              const std::map<InventoryItem, uint8_t>& resource_reward_max,
               float group_reward_pct)
       : GridObjectConfig(type_id, type_name),
         group_name(group_name),
@@ -39,7 +39,7 @@ struct AgentConfig : public GridObjectConfig {
   float action_failure_penalty;
   std::map<InventoryItem, uint8_t> resource_limits;
   std::map<InventoryItem, float> resource_rewards;
-  std::map<InventoryItem, float> resource_reward_max;
+  std::map<InventoryItem, uint8_t> resource_reward_max;
   float group_reward_pct;
 };
 
@@ -54,7 +54,7 @@ public:
   // however, this should not be relied on for correctness.
   std::map<InventoryItem, uint8_t> inventory;
   std::map<InventoryItem, float> resource_rewards;
-  std::map<InventoryItem, float> resource_reward_max;
+  std::map<InventoryItem, uint8_t> resource_reward_max;
   float action_failure_penalty;
   std::string group_name;
   unsigned char color;
@@ -119,7 +119,7 @@ public:
     // TODO: consider doing this only once per step, and not every time the inventory changes.
     float new_reward = 0;
     for (const auto& [item, amount] : this->inventory) {
-      float max_val = static_cast<float>(amount);
+      uint8_t max_val = amount;
       if (this->resource_reward_max.count(item) > 0 && max_val > this->resource_reward_max[item]) {
         max_val = this->resource_reward_max[item];
       }
